Remove custom provider in favour of the react-redux one
